tests working with pytest
