+++ tests/hyperelasticity/test_hyper_elasticity.py
@@ -6,19 +6,19 @@ import meshio
 import os
 import unittest
 
-from jax_am.fem.generate_mesh import Mesh
-from jax_am.fem.models import LinearPoisson
-from jax_am.fem.solver import solver
-from jax_am.fem.utils import modify_vtu_file, save_sol
+from jax_fem.generate_mesh import Mesh
+from jax_fem.models import HyperElasticity
+from jax_fem.solver import solver
+from jax_fem.utils import modify_vtu_file, save_sol
 
 
 class Test(unittest.TestCase):
-    """Test linear Poisson problem
+    """Test hyper-elasticity with cylinder mesh
     """
     def test_solve_problem(self):
         """Compare FEniCSx solution with JAX-FEM
         """
-        problem_name = "linear_poisson"
+        problem_name = "hyperelasticity"
         crt_dir = os.path.dirname(__file__)
         fenicsx_vtu_path_raw = os.path.join(crt_dir, "fenicsx/sol_p0_000000.vtu")
         fenicsx_vtu_path = os.path.join(crt_dir, "fenicsx/sol.vtu")
@@ -27,26 +27,28 @@ class Test(unittest.TestCase):
         cells = fenicsx_vtu.cells_dict['VTK_LAGRANGE_HEXAHEDRON'] # 'hexahedron'
         points = fenicsx_vtu.points
         mesh = Mesh(points, cells)
-        L = 1.
+        H = 10.
 
-        def left(point):
-            return np.isclose(point[0], 0., atol=1e-5)
+        def top(point):
+            return np.isclose(point[2], H, atol=1e-5)
 
-        def right(point):
-            return np.isclose(point[0], L, atol=1e-5)
+        def bottom(point):
+            return np.isclose(point[2], 0., atol=1e-5)
 
-        def dirichlet_val_left(point):
+        def dirichlet_val_bottom(point):
             return 0.
-        
-        def dirichlet_val_right(point):
+
+        def dirichlet_val_top(point):
             return 1.
 
-        location_fns = [left, right]
-        value_fns = [dirichlet_val_left, dirichlet_val_right]
-        vecs = [0, 0]
+        location_fns = [bottom, bottom, bottom, top, top, top]
+        value_fns = [dirichlet_val_bottom, dirichlet_val_bottom, dirichlet_val_bottom,
+                     dirichlet_val_bottom, dirichlet_val_bottom, dirichlet_val_top]
+        vecs = [0, 1, 2, 0, 1, 2]
         dirichlet_bc_info = [location_fns, vecs, value_fns]
 
-        problem = LinearPoisson(mesh, vec=1, dim=3, dirichlet_bc_info=dirichlet_bc_info)
+        problem = HyperElasticity(mesh, vec=3, dim=3, dirichlet_bc_info=dirichlet_bc_info, periodic_bc_info=[[],[],[],[]])
+        problem.p_num_eps = 1e2 # For numerical stability of imposing periodic B.C.
         sol = solver(problem)
 
         jax_vtu_path = os.path.join(crt_dir, "jax_fem/sol.vtu")
@@ -58,6 +60,14 @@ class Test(unittest.TestCase):
 
         print(f"Solution absolute value differs by {np.max(np.absolute(jax_fem_sol - fenicsx_sol))} between FEniCSx and JAX-FEM")
         onptest.assert_array_almost_equal(fenicsx_sol, jax_fem_sol, decimal=5)
+
+        fenicsx_traction = np.load(os.path.join(crt_dir, "fenicsx/traction.npy"))
+        jax_fem_traction = problem.compute_traction(top, sol)[2]
+
+        print(f"FEniCSx computes traction (z-axis) to be {fenicsx_traction}")
+        print(f"JAX-FEM computes traction (z-axis) to be {jax_fem_traction}")
+
+        onptest.assert_almost_equal(fenicsx_traction, jax_fem_traction, decimal=5)
 
 
 if __name__ == '__main__':
